# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/orders": {
      "get": {
        "operationId": "OrdersController_list_v1",
        "parameters": [],
        "responses": { "200": { "description": "" } },
        "security": [{ "itmarket_staff_access": [] }],
        "tags": ["orders"]
      }
    },
    "/api/v1/orders/{id}": {
      "get": {
        "operationId": "OrdersController_get_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "security": [{ "itmarket_staff_access": [] }],
        "tags": ["orders"]
      }
    },
    "/api/v1/orders/{id}/transitions": {
      "post": {
        "operationId": "OrdersController_transition_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TransitionOrderDto" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "security": [{ "itmarket_staff_access": [] }],
        "tags": ["orders"]
      }
    },
...
  },
  "info": {
    "title": "ITMarket API",
    "description": "ITMarket service API",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [{ "url": "/api/v1" }],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "itmarket_customer_session",
        "description": "Opaque, HTTP-only customer session cookie"
      }
    },
    "schemas": {
...
      "TransitionOrderDto": { "type": "object", "properties": {} },
...
    }
  }
}